* Added initial Join, Leave, Schedule, and  Start API Endpoints * Added input and output boundary classes to interact with the application layer * Added all other required models to connect presentation layer with the application layer

+++ src/MeetingScheduler.Domain/Common/IDomainEvent.cs
@@ -3,6 +3,6 @@
 public interface IDomainEvent
 {
     Guid AggregateRootId { get; }
-    int Version { get; }
+    long Version { get; }
     DateTimeOffset Timestamp { get; }
 }
